requirements.txt
discord.py
git+https://github.com/Rishiraj0100/discoutils.git
dpy-appcommands
RistLang==1.1.0
git+https://github.com/Rishiraj0100/shell-reader
pytz
aiomysql
PyNaCl
akinator.py
poetry
aiohttp
asyncpg
sbl.py
lru-dict
parsedatetime
tortoise-orm
jishaku
Flask
pytz
pillow